+++ src/gateway/session-transcript-readers.test.ts
@@ -22,7 +22,6 @@ import {
   readSessionMessageCountAsync,
   readSessionMessagesAsync,
   readSessionMessagesPageWithStatsAsync,
-  readRecentSessionMessagesWithStatsAsync,
   readLatestSessionUsageFromTranscriptAsync,
   type SessionTranscriptReadScope,
 } from "./session-transcript-readers.js";
@@ -201,135 +200,6 @@ describe("session transcript reader facade", () => {
       offset: 0,
       totalMessages: 2,
     });
-  });
-
-  test.each([
-    {
-      name: "compaction",
-      sessionId: "reader-compaction-boundary",
-      markerId: "compaction-boundary",
-      markerKind: "compaction",
-      markerText: "Compaction",
-      events: (sessionId: string) => [
-        { type: "session", version: 3, id: sessionId },
-        {
-          type: "message",
-          id: "before-compaction",
-          parentId: null,
-          message: { role: "user", content: "before compaction" },
-        },
-        {
-          type: "compaction",
-          id: "compaction-boundary",
-          parentId: "before-compaction",
-          timestamp: "2026-08-11T18:00:00.000Z",
-          summary: "summary",
-          firstKeptEntryId: "before-compaction",
-          tokensBefore: 100,
-        },
-        {
-          type: "message",
-          id: "after-compaction",
-          parentId: "compaction-boundary",
-          message: { role: "assistant", content: "after compaction" },
-        },
-      ],
-      expected: ["before compaction", "compaction", "after compaction"],
-    },
-    {
-      name: "reset",
-      sessionId: "reader-reset-boundary",
-      markerId: "reset-boundary",
-      markerKind: "reset",
-      markerText: "Reset",
-      events: (sessionId: string) => [
-        { type: "session", version: 3, id: sessionId },
-        {
-          type: "message",
-          id: "old",
-          parentId: null,
-          message: { role: "user", content: "hidden old turn" },
-        },
-        {
-          type: "message",
-          id: "kept-user",
-          parentId: "old",
-          message: { role: "user", content: "kept question" },
-        },
-        {
-          type: "message",
-          id: "kept-assistant",
-          parentId: "kept-user",
-          message: { role: "assistant", content: "kept answer" },
-        },
-        {
-          type: "reset",
-          id: "reset-boundary",
-          parentId: "kept-assistant",
-          timestamp: "2026-08-11T18:00:00.000Z",
-          reason: "reset",
-          firstKeptEntryId: "kept-user",
-        },
-        {
-          type: "message",
-          id: "post-reset",
-          parentId: "reset-boundary",
-          message: { role: "assistant", content: "new answer" },
-        },
-      ],
-      expected: ["kept question", "kept answer", "reset", "new answer"],
-    },
-  ])("projects $name boundaries through every SQLite history read", async (fixture) => {
-    const scope = await writeTranscript(fixture.sessionId, fixture.events(fixture.sessionId));
-    const summarize = (messages: unknown[]) =>
-      messages.map((message) => {
-        const record = message as { content?: unknown; __openclaw?: { kind?: string } };
-        return record.__openclaw?.kind ?? record.content;
-      });
-
-    const full = await readSessionMessagesAsync(scope, {
-      mode: "full",
-      reason: `${fixture.name} boundary projection test`,
-    });
-    const recent = await readRecentSessionMessagesWithStatsAsync(scope, {
-      maxBytes: 16_384,
-      maxLines: 10,
-      maxMessages: 10,
-    });
-    const markerIndex = fixture.expected.indexOf(fixture.markerKind);
-    const page = await readSessionMessagesPageWithStatsAsync(scope, {
-      maxMessages: 1,
-      offset: fixture.expected.length - markerIndex - 1,
-    });
-    const byId = await readSessionMessageByIdAsync(scope, fixture.markerId);
-    const anchored = await readSessionMessagesAroundIdWithStatsAsync(scope, {
-      messageId: fixture.markerId,
-      maxMessages: 10,
-    });
-
-    expect(summarize(full)).toEqual(fixture.expected);
-    expect(summarize(recent.messages)).toEqual(fixture.expected);
-    expect(recent.totalMessages).toBe(fixture.expected.length);
-    expect(summarize(page.messages)).toEqual([fixture.markerKind]);
-    expect(page.totalMessages).toBe(fixture.expected.length);
-    expect(await readSessionMessageCountAsync(scope)).toBe(fixture.expected.length);
-    expect(byId).toMatchObject({
-      found: true,
-      message: {
-        role: "system",
-        content: [{ type: "text", text: fixture.markerText }],
-        timestamp: Date.parse("2026-08-11T18:00:00.000Z"),
-        __openclaw: {
-          kind: fixture.markerKind,
-          id: fixture.markerId,
-          seq: markerIndex + 1,
-        },
-      },
-      seq: markerIndex + 1,
-    });
-    expect(anchored.found).toBe(true);
-    expect(summarize(anchored.messages)).toEqual(fixture.expected);
-    expect(anchored.totalMessages).toBe(fixture.expected.length);
   });
 
   test("preserves Date.parse semantics for numeric-looking record timestamps", async () => {
